Update UNTIL handling.  iCal exports an UNTIL that is sometimes on the day after the event.  My guess is that it stores the UNTIL internally as 23:59:59 on the last day of the event, and timezone conversion sometimes changes this to the next day.
Luckily, it's possible to recognize this, because iCal uses an illegal
UNTIL format for these anyway -- RFC 2445 says that an UNTIL with a
time MUST be in UTC, and these are not.  Therefore, if an event has
an UNTIL with a time that's not in UTC, and the time is before noon,
subtract a day.  Events with UNTIL= values that comply to the RFC are
not modified by this heuristic.